S-14e extract migrate+fleet to web/modules/migrate.js
Moves loadMigrateAgents, renderMigrateFindings, all migrate step handlers,
and fleet export/dry-run/apply handlers (~360 lines) out of app.js into
migrate.js. initMigrate() collects all top-level event listeners.
escapeHtml is private to the module (was already local to these sections).
Build + 2886/2887 vitest green (known flaky overview-api) + headless smoke:
0 pageerror, 0 404s, migratePage visible, 7 agents loaded, scan+fleet
findings/agent-list markers FOUND.
